+++ src/main/java/com/solarize/solarizeWebBackend/modules/materialUrl/MaterialUrlRepository.java
@@ -2,5 +2,5 @@ package com.solarize.solarizeWebBackend.modules.materialUrl;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MaterialUrlRepository extends JpaRepository<MaterialUrl, Integer> {
+public interface MaterialUrlRepository extends JpaRepository<MaterialUrl, Long> {
 }
